TLV size min

+++ include/TLV.h
@@ -12,9 +12,9 @@ struct iovec *pad_n(uint8_t N);
 struct iovec *hello_short(uint64_t source_id);
 struct iovec *hello_long(uint64_t source_id, uint64_t id);
 struct iovec *neighbour(uint8_t source_ip[16], uint16_t port);
-struct iovec *data(uint64_t sender_id, uint32_t nonce, uint8_t type, uint32_t msg_length, uint8_t *msg);
+struct iovec *data(uint64_t sender_id, uint32_t nonce, uint8_t type, uint8_t msg_length, uint8_t *msg);
 struct iovec *ack(uint64_t sender_id, uint32_t nonce);
-struct iovec *go_away(uint8_t code, uint32_t msg_length, uint8_t *msg);
-struct iovec *warning(uint32_t msg_length, uint8_t *msg);
+struct iovec *go_away(uint8_t code, uint8_t msg_length, uint8_t *msg);
+struct iovec *warning(uint8_t msg_length, uint8_t *msg);
 
 #endif
